requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
streamlit>=1.28.0
opencv-python>=4.8.0
scikit-image>=0.20.0
tqdm>=4.64.0
